requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.29
langchain-core>=0.1.0

# Vector store and embeddings  
chromadb>=0.4.0

# MCP (Model Context Protocol) dependencies
mcp>=0.3.0
langchain-mcp-adapters>=0.1.0

# Knowledge base and RAG
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-docx>=0.8.11

# Integration dependencies
GitPython>=3.1.0
atlassian-python-api>=3.41.0
requests>=2.31.0

# Slack integration
slack-sdk>=3.21.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Async and workflow
asyncio
aiohttp>=3.8.0
celery>=5.3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
